Seperated test files into correct lcoations
