+++ extensions/amp-sidebar/1.0/test/test-amp-sidebar.js
@@ -28,9 +28,6 @@
  /** @const */
  const DEFAULT_TOOLBAR_MEDIA = '(min-width: 768px)';
 
- /** @const */
- const TOOLBAR_CLASS = 'i-amphtml-toolbar';
-
 
  adopt(window);
 
@@ -77,16 +74,9 @@
                  return Promise.resolve();
                });
            // Create our individual toolbars
-           options.toolbars.forEach(toolbar => {
+           options.toolbars.forEach(() => {
              const navToolbar = iframe.doc.createElement('nav');
-             if (toolbar.media) {
-               navToolbar.setAttribute('toolbar', toolbar.media);
-             } else {
-               navToolbar.setAttribute('toolbar', DEFAULT_TOOLBAR_MEDIA);
-             }
-             if (toolbar.toolbarOnlyOnNav) {
-               navToolbar.setAttribute('toolbar-only', 'true');
-             }
+             navToolbar.setAttribute('toolbar', DEFAULT_TOOLBAR_MEDIA);
              const toolbarList = iframe.doc.createElement('ul');
              for (let i = 0; i < 3; i++) {
                const li = iframe.doc.createElement('li');
@@ -115,13 +105,6 @@
            return {iframe, ampSidebar};
          });
        });
-     }
-
-     function resizeIframeToWidth(iframeObject, width, callback) {
-       iframeObject.iframe.setAttribute('width', width);
-       // Force the browser to re-draw
-       iframeObject.win.innerWidth;
-       callback();
      }
 
      beforeEach(() => {
@@ -639,133 +622,6 @@
                 .querySelectorAll('nav[toolbar]'), 0);
          expect(toolbarNavElements.length).to.be.equal(4);
          expect(sidebarElement.implementation_.toolbars_.length).to.be.equal(2);
-       });
-     });
-
-     it('toolbar header should be hidden for a \
-     invalid window size for DEFAULT_TOOLBAR_MEDIA', () => {
-       return getAmpSidebar({
-         toolbars: [true],
-       }).then(obj => {
-         const sidebarElement = obj.ampSidebar;
-         const toolbarElements = Array.prototype
-                .slice.call(sidebarElement.ownerDocument
-                .getElementsByClassName(TOOLBAR_CLASS), 0);
-         resizeIframeToWidth(obj.iframe, '1px', () => {
-           expect(toolbarElements.length).to.be.above(0);
-           sidebarElement.implementation_.toolbars_.forEach(toolbar => {
-             toolbar.onLayoutChange();
-           });
-           expect(toolbarElements[0].parentElement.style.display)
-               .to.be.equal('none');
-         });
-       });
-     });
-
-     it('toolbar header should be shown for a \
-     valid window size for DEFAULT_TOOLBAR_MEDIA', () => {
-       return getAmpSidebar({
-         toolbars: [true],
-       }).then(obj => {
-         const sidebarElement = obj.ampSidebar;
-         const toolbarElements = Array.prototype
-                .slice.call(sidebarElement.ownerDocument
-                .getElementsByClassName(TOOLBAR_CLASS), 0);
-         resizeIframeToWidth(obj.iframe, '4000px', () => {
-           expect(toolbarElements.length).to.be.above(0);
-           sidebarElement.implementation_.toolbars_.forEach(toolbar => {
-             toolbar.onLayoutChange();
-           });
-           expect(toolbarElements[0].parentElement.style.display)
-               .to.be.equal('');
-         });
-       });
-     });
-
-     it('should hide <nav toolbar> elements with toolbar-only, \
-     inside the sidebar, but not inside the toolbar, for a valid \
-     window size for DEFAULT_TOOLBAR_MEDIA', () => {
-       return getAmpSidebar({
-         toolbars: [{
-           toolbarOnlyOnNav: true,
-         }],
-       }).then(obj => {
-         const sidebarElement = obj.ampSidebar;
-         const toolbars = sidebarElement.implementation_.toolbars_;
-         resizeIframeToWidth(obj.iframe, '4000px', () => {
-           toolbars.forEach(toolbar => {
-             toolbar.onLayoutChange();
-           });
-           const toolbarNavElements = Array.prototype
-                  .slice.call(sidebarElement.ownerDocument
-                  .querySelectorAll('nav[toolbar]'), 0);
-           const hiddenToolbarNavElements = Array.prototype
-                  .slice.call(sidebarElement.ownerDocument
-                  .querySelectorAll('nav[style]'), 0);
-           expect(toolbarNavElements.length).to.be.equal(2);
-           expect(hiddenToolbarNavElements.length).to.be.equal(1);
-           expect(toolbars.length).to.be.equal(1);
-         });
-       });
-     });
-
-     it('toolbar should be in the hidden state \
-     when it is not being displayed', () => {
-       return getAmpSidebar({
-         toolbars: [true],
-       }).then(obj => {
-         const toolbars = obj.ampSidebar.implementation_.toolbars_;
-         resizeIframeToWidth(obj.iframe, '1px', () => {
-           toolbars.forEach(toolbar => {
-             toolbar.onLayoutChange();
-             expect(toolbar.isToolbarShown_()).to.be.false;
-           });
-         });
-       });
-     });
-
-     it('toolbar should be in the shown state \
-     when it is being displayed', () => {
-       return getAmpSidebar({
-         toolbars: [true],
-       }).then(obj => {
-         const toolbars = obj.ampSidebar.implementation_.toolbars_;
-         resizeIframeToWidth(obj.iframe, '4000px', () => {
-           toolbars.forEach(toolbar => {
-             toolbar.onLayoutChange();
-             expect(toolbar.isToolbarShown_()).to.be.true;
-           });
-         });
-       });
-     });
-
-     it('toolbar should not be able to be shown \
-     if already in the shown state', () => {
-       return getAmpSidebar({
-         toolbars: [true],
-       }).then(obj => {
-         const toolbars = obj.ampSidebar.implementation_.toolbars_;
-         resizeIframeToWidth(obj.iframe, '4000px', () => {
-           toolbars.forEach(toolbar => {
-             toolbar.onLayoutChange();
-             expect(toolbar.attemptShow_()).to.be.undefined;
-           });
-         });
-       });
-     });
-
-     it('toolbar should be able to be shown \
-     if not in the shown state, and return a promise', () => {
-       return getAmpSidebar({
-         toolbars: [true],
-       }).then(obj => {
-         const toolbars = obj.ampSidebar.implementation_.toolbars_;
-         resizeIframeToWidth(obj.iframe, '1px', () => {
-           toolbars.forEach(toolbar => {
-             toolbar.onLayoutChange();
-             expect(toolbar).to.exist;
-           });
-         });
        });
      });
    });
